Adding fallback if users file doesn't exist. Removing an unhittable code block
Starting 2ee8dfc, conf.users could no longer be null/undefined, and the
code block in the posthook's user resolver would never be hit. This
change accommodates a base use case where the users config property is
not set and where users.json does not exist. In that scenario, the
server will defer to the global token and key instead of failing a
couple lines later when it does a type check on the parsed people
object.